Trigger startup configuration event consistently
Before this commit, the startup event `wwu_setup.3`, which currently
configures the diplomatic options, would not trigger at start. Only when
the game was reloaded or somehow reset this event would trigger.

The cause of this error seems to be the order the country flags are set
and read. `on_startup` did set the country flag `host_player`, but the
values are not read properly if the event that requires this flag is
also fired at the same time.

Instead of using `host_player` as the flag for determining the host, the
event now uses `event_target:executor`. And by also triggering both
`wwu_setup.1` (which configures the world) and `wwu_setup.3` in
sequence, `wwu_setup.3` now always triggers consistently 1 day after game
start.