Conform to YAML syntax
This commit fixes a lot of invalid yaml syntax. A common theme is
incorrect quotes inside a quoted string.

Localized files have also been updated as part of this work, so all
files touched by this commit are now identical with the english version,
apart from the top level language key.

This work is required in order to add more advanced features, such as
spellchecking or machine translation.

The following commit adds code to verify that almost all `*.yml` files
in `localisation/` are conforming to YAML syntax.