Spellchecker for English localisation
This commit adds a spellchecker to verify that the English descriptions,
titles, etc are not misspelled. `cspell` is able to understand quite a
lot of machine friendly syntax, so the rate of false-positives are
pretty low.

The spellchecking is run automatically when pushing code through Github
Actions. To run spellchecking locally, take a look at
`# Development/tools/spell-checking/readme.md`.

An example of an error message you could get can be seen below:

```
root@532d19c38bfe:/host_dir# python3 \#\ Development/tools/spell-checking/spellcheck.py
/host_dir/localisation/WWU_disasters_l_english.yml:46:76 - Unknown word (siezed)
Spell checking failed
```

Since there are so many words to keep track of, a few specialized
dictionaries has been added so names, places, cultures, artifacts, eu4
terminology, etc can be separated. It is easy to create a new dictionary
if needed.

It is outside the scope of this commit to categories all of the words in
added to `cspell.json`. The following commits will show how to fix a few
spelling mistakes. But in general the workflow is to delete the word
from the `words` list in `cspell.json` and run `spellcheck.py`, either
through Github Actions or locally.