Fix refactor regression
Commit b93059ca7049928f64ea1a6f7aac6e01593bd723 introduced a regression,
caused by the changes in the number of sub-directories to the validation
script.

This commit fixes this regression and adds a test in the validation code
so that there always should be at least 1 localisation file found when
globbing for them.

Also added a missing `"` that were added since the regression were
introduced.