Delete unused elder scrolls code
The code deleted in this commit was found by searching for elder scrolls
specific names, and then checking if the relevant code block were
referenced in the code anywhere.

A few lines in `esu_*` and `_es_*` files are still in use, but that code
has drowned in unused code which references places and characters not
relevant for this universe.