Delete unused code and correct spelling
Follow-up to commit 2f0619b63e68fc88d4f0b64bfa12a005539f1aca, which
deleted some unused code.

This commit deletes unused code and reworks the description of `tobacco`
trade good.

The reason why the description of the trade good `tobacco` required a
rework is that it contained references to a unrelated universe.

Most of the code was deleted by simply removing the relevant entry in
`cspell.json` and running the spell checking code.