Verify event and mission modifiers are unused
While commit 80ded86857b2056f659f88efcf1e3bf6b8884aa4 deleted all unused
event and mission modifiers, it does not prevent other unused modifiers
to appear. This commit on the other hand, enables a very naive system
for detecting unused event and mission modifiers.

In short, the system tokenizes and creates a parse tree of the modifier
files. After that it takes the id of the modifier and does a search in
all `.txt` and `.yml` files outside of `common/event_modifiers` to try
to find *any* match in the raw text read.
If there is any match, the script will assume that the modifier is in
use and move to the next id to check.
If there is no match, then the script will fail, and print that it could
not find the modifier in question, and therefore should probably be
deleted.

An obvious weakness to this system is partial matches, for instance if a
modifier id is just a subset of another used modifier id then the script
will always think the modifier is in use. For example `delayed_payments`
will always be in use if there is another modifier `delayed_payments_2`
that is *actually* in use.

To mitigate those issues would likely require complete parsing of the
eu4 modding syntax, which is outside the scope of this commit.

`lark` was chosen as the library to use to create the parse tree, which
seems to be one the most modern of tokenizer and parse tree libraries
for python.

Added a Github Action job to automatically run the script when
triggered.