Wowpedia titles as vocabulary
This commit adds a dictionary that contains a filtered selection of the
titles of each page from `wowpedia.fandom.com`. Also added a script to
easily update the dictionary.

The dictionary entries are not perfectly filtered by any means, but
should be good enough to reduce the number of words to maintain manually
maintained cspell dictionaries.

The logic of the update script is fairly straightforward, download a
compressed version of `wowpedia`, filter away some groups of unrelated
pages, such as forum discussions and user pages.

Then get the title for each of the relevant pages and split it into
special characters such as `:` and `/`, before splitting the titles into
individual words. The words themselves also need to be carefully
sanitized as well, suffixes such as `,`, `'s`, `.`, `!`, etc should be
removed, and prefixes such as `(` and `"`, etc should also be removed
before finally writing the word list to file.

This commit also removes duplicate entries that have already been
defined by `wowpedia`, which is why the `words` list in `cspell.json`
and the other dictionaries have deleted entries.