Delete unused generic event text
None of the `EVTNAME`, `EVTDESC`, `EVTOPT`, `EVTDESHOR`, `EVTNAMEHR` and
`EVTOPTHOR` keys were used anywhere else in the repo apart from
`localisation/replace/generic_events_l_*.yml`, which indicates that it
should be okay to delete them.

Due to the fact that `generic_events_l_*.yml` does not contain valid
`YAML`, a script that parses the file and returns the relevant key had
to be created. This is of course more error prone than using a proper
`YAML` parsing library.
The script used to find the unused keys has not been included in this
commit, since it was intended to just run it once.

Additionally, this commit also deletes some event modifiers since there
are no longer any references to them anymore, such as `obscurantism`.

Also cleaned up words in `cspell.json` as they were only used in the
text deleted by this commit. If some of them are valid words they can
always be added back when needed.