Make localization files yaml compliant
According to the EU4 wiki,
https://eu4.paradoxwikis.com/Localisation#Format the number after the
colon after the localization keys in `l_*.yml` localization files
indicates the version of the string and is optional.

Since this mod is managed by `Git` this version field is not needed by
WWU and can therefore be deleted, which is required to be `YAML`
compliant.

This commit also removes some redundant `"` that were detected after the
version numbers were removed.

The `YAML` validation logic that ignored those files has also been
deleted.