Replace pyyaml with pdxloc for file validation
`pdxloc` is a library I created for handling localization files, which
at the current moment can read/write localization files to disk.

As this library uses `ruamel.yaml` instead of `pyyaml` as part of the
logic to read/write localization files, this is basically what this
commit does.

The Ubuntu image had to be bumped to 22.04 as `pdxloc` only supports
versions newer or equal to python 3.9. In this case the version to use
to test this code was set to 3.10.